+++ RELEASES.md
@@ -1,3 +1,7 @@
+# 0.1.5
+ * More robust dead connection logic.
+ * Fixed issues with input cursor and ctrl + c
+
 # 0.1.4
  * Centralize completer logic.